renamed home routes to landing routes, added error routes js
