+++ controllers/errorRoutes.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 
 router.get('/', async(req, res) => {
-    res.render('login');
+    res.render('error');
 })
 
 module.exports = router;
