+++ controllers/index.js
@@ -1,10 +1,10 @@
 const router = require('express').Router();
 
 const apiRoutes = require('./api');
-const homeRoutes = require('./homeRoutes');
+const landingRoutes = require('./landingRoutes');
 const loginRoutes = require('./loginRoutes');
 
-router.use('/', homeRoutes);
+router.use('/', landingRoutes);
 router.use('/api', apiRoutes);
 router.use('/login', loginRoutes);
